requirements.txt
# Minimal requirements for running the langgraph chatbot project
# Pin versions as appropriate for your environment. Update if you use different libraries.
langchain>=0.0.200
openai>=0.27.0
langgraph>=0.1.0
fastapi>=0.95.0
uvicorn>=0.22.0
requests>=2.31.0
python-dotenv>=1.0.0
azure-cosmos>=4.3.0
azure-identity>=1.12.0
pydantic>=1.10.0

# Optional / dev
pytest>=7.0.0
flake8>=6.0.0
azure-cosmos
langchain
langchain-community
langchain-experimental
fastapi
python-dotenv
streamlit
uvicorn
pydantic
httpx
requests
langchain_google_community
langchain_tavily
langchain_groq
langchain_openai
langgraph
langchain-google-community[places]
####-e .